fix(Dependencies): Upgrades mongodb driver to ^4.7.0 (LLC-1690) (#314)
BREAKING CHANGE: when using this repo be mindful we now use official MongoDB TS Types which are packed with the mongodb driver v4.0.0, so if you're using a version of the mongodb driver that's lower than v4 you will likely be using the mongodb community types @types/mongodb which could clash with the shape of objects now defined in this project (i.e the official MongoDB types). To resolve, you'd likely need to upgrade to use mongodb driver ^4.0.0.